Refactor stories
